third_party/WebKit/LayoutTests/inspector/console/console-trace-in-eval-expected.txt

CONSOLE MESSAGE: line 5: console.trace
Tests that when console.trace is called in eval'ed script ending with //# sourceURL=url it will dump a stack trace that will have the url as the script source. Bug 47252.

evalURL.js:5 console.traceb @ evalURL.js:5a @ evalURL.js:10evalSource @ evalURL.js:13(anonymous) @ evalURL.js:14(anonymous) @ console-trace-in-eval.html:26
